Split pop into poor, middle and rich (for now), as it's now possible to them to be paid separately. Added service job to transfert money from rich to poor. many bug fixes industryneed like popneed. Simplify need interface. refined the povincegood price algo: now use consolidated data and not unreliable data. add helper methods to needs and industry todo: create upgrade/downgrade of pop.

+++ src/remi/ssp/economy/Job.java
@@ -2,7 +2,7 @@ package remi.ssp.economy;
 
 public interface Job {
 
-	public long getPreviousSalary();
+	public double getPreviousSalary();
 //	public int getNbWorkers(); //stored in pop
 //	public void addWorkers(int nb);
